+++ app/src/main/java/nextstep/signup/MainActivity.kt
@@ -18,21 +18,15 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import nextstep.signup.model.Email
-import nextstep.signup.model.EmailValidResult
 import nextstep.signup.model.Password
 import nextstep.signup.model.PasswordConfirm
-import nextstep.signup.model.PasswordConfirmValidResult
-import nextstep.signup.model.PasswordValidResult
 import nextstep.signup.model.User
 import nextstep.signup.model.UserName
-import nextstep.signup.model.UserNameValidResult
 import nextstep.signup.ui.component.ButtonComponent
 import nextstep.signup.ui.component.TextComponent
-import nextstep.signup.ui.component.TextFieldComponent
 import nextstep.signup.ui.theme.SignupTheme
 
 class MainActivity : ComponentActivity() {
@@ -56,7 +50,7 @@ class MainActivity : ComponentActivity() {
 }
 
 @Composable
-fun SignUpScreen() {
+private fun SignUpScreen() {
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
@@ -103,112 +97,6 @@ private fun SignUpInputBox(
     )
 }
 
-@Composable
-fun UserNameComposable(
-    userName: UserName,
-    onUserNameChange: (String) -> Unit,
-) {
-    TextFieldComponent(
-        newValue = userName.userName,
-        onValueChange = onUserNameChange,
-        label = stringResource(R.string.main_user_name),
-        supportingText = {
-            val errorMessage = userName.getErrorMessage() ?: return@TextFieldComponent
-            TextComponent(description = errorMessage)
-        },
-        isError = userName.isError(),
-    )
-    Spacer(modifier = Modifier.size(36.dp))
-}
-
-private fun UserName.getErrorMessage(): String? =
-    when (this.validate()) {
-        UserNameValidResult.Blank -> null
-        UserNameValidResult.InvalidLength -> USERNAME_LENGTH_ERROR
-        UserNameValidResult.InvalidCharacter -> USERNAME_FORM_ERROR
-        else -> null
-    }
-
-@Composable
-fun EmailComposable(
-    email: Email,
-    onEmailChange: (String) -> Unit,
-) {
-    TextFieldComponent(
-        newValue = email.email,
-        onValueChange = onEmailChange,
-        label = stringResource(R.string.main_email),
-        supportingText = {
-            val errorMessage = email.getErrorMessage() ?: return@TextFieldComponent
-            TextComponent(description = errorMessage)
-        },
-        isError = email.isError(),
-        keyboardType = KeyboardType.Email,
-    )
-    Spacer(modifier = Modifier.size(36.dp))
-}
-
-private fun Email.getErrorMessage(): String? =
-    when (this.validate()) {
-        EmailValidResult.Blank -> null
-        EmailValidResult.InvalidForm -> EMAIL_FORM_ERROR
-        else -> null
-    }
-
-@Composable
-fun PasswordComposable(
-    password: Password,
-    onPasswordChange: (String) -> Unit,
-) {
-    TextFieldComponent(
-        newValue = password.password,
-        onValueChange = onPasswordChange,
-        label = stringResource(R.string.main_password),
-        supportingText = {
-            val errorMessage = password.getErrorMessage() ?: return@TextFieldComponent
-            TextComponent(description = errorMessage)
-        },
-        isError = password.isError(),
-        keyboardType = KeyboardType.Password,
-    )
-    Spacer(modifier = Modifier.size(36.dp))
-}
-
-private fun Password.getErrorMessage(): String? =
-    when (this.validate()) {
-        PasswordValidResult.Blank -> null
-        PasswordValidResult.InvalidLength -> PASSWORD_LENGTH_ERROR
-        PasswordValidResult.InvalidCharacter -> PASSWORD_FORM_ERROR
-        else -> null
-    }
-
-@Composable
-fun PasswordConfirmComposable(
-    password: Password,
-    passwordConfirm: PasswordConfirm,
-    onPasswordConfirmChange: (String) -> Unit,
-) {
-    TextFieldComponent(
-        newValue = passwordConfirm.passwordConfirm,
-        onValueChange = onPasswordConfirmChange,
-        label = stringResource(R.string.main_password_confirm),
-        supportingText = {
-            val errorMessage = passwordConfirm.getErrorMessage(password.password) ?: return@TextFieldComponent
-            TextComponent(description = errorMessage)
-        },
-        isError = passwordConfirm.isError(password.password),
-        keyboardType = KeyboardType.Password,
-    )
-    Spacer(modifier = Modifier.size(42.dp))
-}
-
-private fun PasswordConfirm.getErrorMessage(password: String): String? =
-    when (this.validate(password)) {
-        PasswordConfirmValidResult.Blank -> null
-        PasswordConfirmValidResult.Invalid -> PASSWORD_CONFIRM_ERROR
-        else -> null
-    }
-
 @Preview(showBackground = true)
 @Composable
 private fun GreetingPreview() {
@@ -216,10 +104,3 @@ private fun GreetingPreview() {
         SignUpScreen()
     }
 }
-
-private const val USERNAME_LENGTH_ERROR = "이름은 2~5자여야 합니다."
-private const val USERNAME_FORM_ERROR = "이름에는 숫자나 기호가 포함될 수 없습니다."
-private const val EMAIL_FORM_ERROR = "이메일 형식이 올바르지 않습니다."
-private const val PASSWORD_LENGTH_ERROR = "비밀번호는 8~16자여야 합니다."
-private const val PASSWORD_FORM_ERROR = "비밀번호는 영문과 숫자를 포함해야 합니다."
-private const val PASSWORD_CONFIRM_ERROR = "비밀번호가 일치하지 않습니다."
